make machine stuff abstract sets move Card::Machine to Card::Set::Abstract::Machine move Card::MachineInput to Card::Set::Abstract::MachineInput
