+++ card/mod/machines/set/type/skin.rb
@@ -1,5 +1,5 @@
-include Machine
-include MachineInput
+include_set Abstract::Machine
+include_set Abstract::MachineInput
 
 store_machine_output filetype: "css"
