+++ card/mod/settings/set/right/style.rb
@@ -1,5 +1,5 @@
 require "sass"
-include Machine
+include_set Abstract::Machine
 
 store_machine_output filetype: "css"
